requirements.txt
discord.py
requests
selenium
python-dotenv